+++ ncollide_pipeline/narrow_phase/ball_ball.rs
@@ -66,7 +66,7 @@ impl<N, P, V, M> CollisionDetector<N, P, V, M> for BallBall<N, P, V, M>
     }
 
     #[inline]
-    fn num_colls(&self) -> uint {
+    fn num_colls(&self) -> usize {
         match self.contact {
             None    => 0,
             Some(_) => 1
