+++ ncollide_pipeline/world/collision_groups.rs
@@ -24,7 +24,7 @@ impl CollisionGroups {
     ///
     /// This does not enable self-collision.
     #[inline]
-    pub fn new_with_groups(groups: &[uint]) -> CollisionGroups {
+    pub fn new_with_groups(groups: &[usize]) -> CollisionGroups {
         let mut res = CollisionGroups { tag: 0 };
 
         for group in groups.iter() {
@@ -39,7 +39,7 @@ impl CollisionGroups {
     ///
     /// This does not enable self-collision.
     #[inline]
-    pub fn new_without_groups(groups: &[uint]) -> CollisionGroups {
+    pub fn new_without_groups(groups: &[usize]) -> CollisionGroups {
         let mut res = CollisionGroups { tag: ALL_GROUPS };
 
         for group in groups.iter() {
@@ -67,7 +67,7 @@ impl CollisionGroups {
 
     /// Enable collision detection with a specific group.
     #[inline]
-    pub fn enable_collisions_with_group(&mut self, group_id: uint) {
+    pub fn enable_collisions_with_group(&mut self, group_id: usize) {
         assert!(group_id < 30, "There are at most 30 groups indexed from 0 to 29 (included).");
 
         self.tag = self.tag | (1 << group_id);
@@ -75,7 +75,7 @@ impl CollisionGroups {
 
     /// Disables collision detection with a specific group.
     #[inline]
-    pub fn disable_collisions_with_group(&mut self, group_id: uint) {
+    pub fn disable_collisions_with_group(&mut self, group_id: usize) {
         assert!(group_id < 30, "There are at most 30 groups indexed from 0 to 29 (included).");
 
         self.tag = self.tag & !(1 << group_id);
@@ -95,7 +95,7 @@ impl CollisionGroups {
 
     /// Tests whether collisions with a given group is enabled.
     #[inline]
-    pub fn can_collide_with_group(&self, group_id: uint) -> bool {
+    pub fn can_collide_with_group(&self, group_id: usize) -> bool {
         assert!(group_id < 30, "There are at most 30 groups indexed from 0 to 29 (included).");
         self.tag & (1 << group_id) != 0
     }
